chore: setup nexe-fs as a standalone shim
