+++ package.json
@@ -2,7 +2,7 @@
   "name": "nexe",
   "description": "Create a single executable out of your Node.js application",
   "license": "MIT",
-  "version": "2.0.0-next.12",
+  "version": "2.0.0-next.13",
   "contributors": [
     "Craig Condon <craig.j.condon@gmail.com> (http://crcn.io)",
     "Jared Allard <jaredallard@outlook.com>",
@@ -36,20 +36,22 @@
     "app-builder": "^5.2.0",
     "caw": "^2.0.1",
     "chalk": "^2.4.1",
+    "cherow": "^1.6.8",
     "download": "^7.0.0",
     "globby": "^8.0.1",
     "got": "^8.3.1",
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
     "ora": "^2.1.0",
-    "pify": "^3.0.0",
-    "resolve-dependencies": "^2.0.0",
+    "pify": "^4.0.0",
+    "resolve-dependencies": "^2.1.1",
     "rimraf": "^2.6.2"
   },
   "devDependencies": {
     "@types/chai": "^4.1.3",
     "@types/execa": "^0.9.0",
-    "@types/globby": "^6.1.0",
+    "@types/globby": "^8.0.0",
+    "@types/got": "^8.3.3",
     "@types/minimist": "^1.2.0",
     "@types/mkdirp": "^0.5.2",
     "@types/mocha": "^5.2.0",
@@ -60,11 +62,11 @@
     "chai": "^4.1.2",
     "execa": "^0.10.0",
     "mocha": "^5.1.1",
-    "prettier": "^1.12.1",
-    "ts-node": "6.0.3",
+    "prettier": "^1.14.2",
+    "ts-node": "^7.0.0",
     "tslint": "^5.10.0",
     "tslint-config-prettier": "^1.12.0",
     "tslint-plugin-prettier": "^1.3.0",
-    "typescript": "2.8.3"
+    "typescript": "^3.0.1"
   }
 }
