+++ src/patches/boot-nexe.ts
@@ -1,7 +1,7 @@
-const fs = require('fs')
-const fd = fs.openSync(process.execPath, 'r')
-const stat = fs.statSync(process.execPath)
-const footer = Buffer.alloc(32, 0)
+const fs = require('fs'),
+  fd = fs.openSync(process.execPath, 'r'),
+  stat = fs.statSync(process.execPath),
+  footer = Buffer.alloc(32, 0)
 
 fs.readSync(fd, footer, 0, 32, stat.size - 32)
 
@@ -9,10 +9,10 @@ if (!footer.slice(0, 16).equals(Buffer.from('<nexe~~sentinel>'))) {
   throw 'Invalid Nexe binary'
 }
 
-const contentSize = footer.readDoubleLE(16)
-const resourceSize = footer.readDoubleLE(24)
-const contentStart = stat.size - 32 - resourceSize - contentSize
-const resourceStart = contentStart + contentSize
+const contentSize = footer.readDoubleLE(16),
+  resourceSize = footer.readDoubleLE(24),
+  contentStart = stat.size - 32 - resourceSize - contentSize,
+  resourceStart = contentStart + contentSize
 
 Object.defineProperty(
   process,
@@ -25,9 +25,10 @@ Object.defineProperty(
       },
       set: function(value: any) {
         if (nexeHeader) {
-          throw new Error('__nexe cannot be reconfigured')
+          throw new Error('This property is readonly')
         }
         nexeHeader = Object.assign({}, value, {
+          blobPath: process.execPath,
           layout: {
             stat,
             contentSize,
@@ -44,10 +45,10 @@ Object.defineProperty(
   })()
 )
 
-const contentBuffer = Buffer.from(Array(contentSize))
+const contentBuffer = Buffer.from(Array(contentSize)),
+  Module = require('module')
+
 fs.readSync(fd, contentBuffer, 0, contentSize, contentStart)
 fs.closeSync(fd)
-const Module = require('module')
-const mod = new Module(process.execPath, null)
-mod.loaded = true
-mod._compile(contentBuffer.toString(), process.execPath)
+
+new Module(process.execPath, null)._compile(contentBuffer.toString(), process.execPath)
